requirements.txt
asgiref==3.5.2
attrs==21.4.0
backports.zoneinfo==0.2.1
certifi==2022.6.15
charset-normalizer==2.1.0
coverage==6.4.2
dj-database-url==0.5.0
Django==4.0.6
django-cors-headers==3.13.0
django-filter==22.1
django-on-heroku==1.1.2
django-phonenumber-field==6.3.0
django-phonenumbers==1.0.1
djangorestframework==3.13.1
djangorestframework-simplejwt==5.2.0
gunicorn==20.1.0
idna==3.3
importlib-metadata==4.12.0
iniconfig==1.1.1
Markdown==3.3.7
packaging==21.3
phonenumbers==8.12.51
Pillow==9.2.0
pluggy==1.0.0
psycopg2-binary==2.9.3
py==1.11.0
PyJWT==2.4.0
pyparsing==3.0.9
pytest==7.1.2
pytest-django==4.5.2
python-dotenv==0.20.0
pytz==2022.1
razorpay==1.3.0
requests==2.28.1
sqlparse==0.4.2
tomli==2.0.1
urllib3==1.26.10
whitenoise==6.2.0
zipp==3.8.1
